Refactor tests and moved fixture to conftest
added negative tests and refactor valid_teams
refactor tests for different files and fix user validation
added validation before changing the status and priority
added parametrize
