+++ falcon/test_support.py
@@ -9,6 +9,10 @@ class Random:
         return ''.join(random.choices(string.ascii_letters + string.digits, k=length))
 
     @ staticmethod
+    def alphaOnly(length: int = 5):
+        return ''.join(random.choices(string.ascii_letters, k=length))
+
+    @ staticmethod
     def email():
         return Random.string(10)+"@redhat.com"
